Huge General cleanup
