+++ Expenses/Assets.xcassets/AppIcon.appiconset/Contents.json
