+++ Expenses/Assets.xcassets/AppIcon.appiconset/Icon-Notification.png
