+++ Expenses/Assets.xcassets/AppIcon.appiconset/Icon-Small-40.png
