+++ Expenses/Assets.xcassets/AppIcon.appiconset/Icon-Small-40@2x-1.png
