+++ Expenses/Assets.xcassets/AppIcon.appiconset/Icon-Small@3x.png
