+++ Expenses/Assets.xcassets/AppIcon.appiconset/edited.png
