+++ Expenses/Assets.xcassets/AppIcon.appiconset/icon_76@2x.png
