+++ Expenses/Assets.xcassets/AppIcon.appiconset/icon_83.5@2x.png
