+++ Expenses/Assets.xcassets/Orange/Contents.json
