+++ Expenses/Assets.xcassets/Red/Contents.json
