+++ Expenses/Assets.xcassets/Red/PrimaryRed.colorset/Contents.json
