+++ Expenses/Assets.xcassets/Red/SecondaryRed.colorset/Contents.json
