+++ Expenses/Views/GroupBoxPickersView.swift
@@ -9,7 +9,7 @@ import SwiftUI
 
 struct GroupBoxPickersView: View {
     
-    @EnvironmentObject var settings: AppSettings
+    @EnvironmentObject var settings: AppSettingsViewModel
     
     //TODO: Have the buttons disbales if categories and banks are empty.
