+++ Expenses/Views/HomeExtensions.swift
@@ -1,0 +1,21 @@
+//
+//  HomeExtensions.swift
+//  Expenses
+//
+//  Created by Anfernee Viduya on 4/14/22.
+//
+
+import Foundation
+import SwiftUI
+
+//MARK: Extracted Views of HomeView
+
+extension HomeView {
+    
+    //MARK: List view of 5 Recent transactions.
+    
+
+}
+
+
+
